Updated Products Page
Example 1 for Great Lakes Blend

Will be creating product-2.html and product-2.css for other product. Can use the dropdown menus to link to next page.

Is there a way that we can toggle through the products on one page rather than having to send the user to another link?I will look into that.